requirements.txt
flask
botbuilder-core
openai
azure-cognitiveservices-speech
azure-cognitiveservices-vision-computervision
